NonlinearBandits

Documentation for NonlinearBandits.

NonlinearBandits.BayesLM

NonlinearBandits.BayesPM

NonlinearBandits.Index

NonlinearBandits.expand

NonlinearBandits.fit!

NonlinearBandits.std

NonlinearBandits.tpbasis

NonlinearBandits.BayesLM — Method

BayesLM(d::Int; <keyword arguments>)

Construct a Bayesian linear model.

Arguments

λ::tFloat64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

NonlinearBandits.BayesPM — Method

BayesPM(basis::Vector{Index}, limits::AbstractMatrix; λ::Float64=1.0,
        shape0::Float64=1e-3, scale0::Float64=1e-3)

Construct a Bayesian linear model on polynomial features.

Arguments

basis::Vector{Index}: Vector of monomial indices.

limits::AbstractMatrix: Matrix with two columns defining the lower/upper limits of the space.

λ::Float64=1.0: Prior covariance scale factor.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e=3: Inverse-gamma prior scale hyperparameter.

NonlinearBandits.Index — Type

Index(dim::Vector{Int64}, deg::Vector{Int64})

Multivariate monomial index.

The monomial x[1] * x[3]^2can be encoded usingdim = [1, 3],deg = [1, 2]`

NonlinearBandits.expand — Method

expand(X::AbstractMatrix, basis::Vector{Index}, limits::AbstractMatrix;
       J::Union{Nothing,Int64}=nothing)

Expand the columns of X into a rescaled legendre polynomial basis.

Arguments

X::AbstractMatrix: Matrix with observations stored as columns.

basis::Vector{<:Index}: Vector of monomial indices.

limits::AbstractMatrix: Matrix with two columns defining the lower/upper limits of the space.

J::Union{Nothing, Int64}=nothing: The maximum degree of the basis. Inferred if not specified.

NonlinearBandits.fit! — Method

fit!(model, X::AbstractMatrix, y::AbstractMatrix)

Update the parameters of model.

Arguments

X::AbstractMatrix: A matrix with observations stored as columns.

y::AbstractMatrix: A matrix with 1 row of response variables.

NonlinearBandits.std — Method

std(lm::BayesLM)

Computed the posterior expected standard deviation from the model lm.

NonlinearBandits.tpbasis — Method

tpbasis(d::Int64, J::Int64)

Construct the d-dimensional truncated tensor-product basis.

All index terms have a degree ≤ J.

See also Index