Bandits API

NonlinearBandits.AbstractContextSampler — Type

Called with an integer n > 0 to genrate a d x n matrix of contexts.

NonlinearBandits.AbstractDriver — Type

Driver to manage how the bandit policy interacts with its environment.

NonlinearBandits.AbstractMetric — Type

A metric that will be called as metric(X, a, r) after each batch outputted by a driver.

NonlinearBandits.AbstractPolicy — Type

A callable policy that outputs actions via policy(X) and is updated via update!

NonlinearBandits.AbstractRewardSampler — Type

A callable object to ouput rewards via sampler(X, a).

NonlinearBandits.BanditDataset — Type

BanditDataset(d::Int64)

Stores the trajectory of a bandit simulation.

NonlinearBandits.append_data! — Method

append_data!(data::BanditDataset, X::AbstractMatrix, a::AbstractVector{<:Int},
          r::AbstractMatrix)

Add a batch of data to the dataset.

NonlinearBandits.arm_data — Method

rm_data(data::BanditDataset, a::Int64)

Return the data associated with arm a.

Context Samplers

NonlinearBandits.UniformContexts — Type

UniformContexts(limits::Matrix{Float64})

Construct a callable object to generate uniform contexts.

Policies

NonlinearBandits.update! — Function

update!(pol::AbstractPolicy, X::AbstractMatrix, a::AbstractVector{<:Int}, 
        r::AbstractMatrix)

Update pol with a batch of data.

NonlinearBandits.RandomPolicy — Type

RandomPolicy(num_actions::Int64)

Construct a policy that chooses actions at random.

NonlinearBandits.PolynomialThompsonSampling — Type

PolynomialThompsonSampling(limits::Matrix{Float64}, num_arms::Int64, 
                           initial_batches::Int64, retrain_freq::Int64;
                           <keyword arguments>)

Construct a Thompson sampling policy that uses a PartitionedBayesPM to model the expected rewards.

Arguments

limits::Matrix{Float64}: Matrix with two columns defining the lower/upper limits of the space.

num_arms::Int64: The number of available actions.

inital_batches::Int64: The number of batches to sample before training the polnomial models.

retrain_freq::Int64: The frequency (in terms of batches) at which the partition/basis selection is retrained from scratch.

α::Float64=1.0: Thompson sampling inflation. α > 1 and increasing alpha increases the amount of exploration.

Jmax::Int64=3: The maximum degree of any polynomial region.

Pmax::Int64=100: The maximum number of features in any polynomial region.

Kmax::Int64=500: The maximum number of regions in the partition.

λ::Float64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

ratio::Float64=1.0: Polynomial degrees are reduced until size(X, 2) < ratio * length(tpbasis(d, J)).

tol::Float64=1e-4: The required increase in the model evidence to accept a split.

verbose_retrain::Bool=true: Print details of the partition search.

NonlinearBandits.NeuralEncoder — Type

NeuralEncoder(d_int::Int64, d_out::Int64, layer_sizes::Vector{Int64})

Construct a callable object that returns the final layer activations of a neural network (which can be trained with a contextual multi armed bandit trajectory).

NonlinearBandits.NeuralLinear — Type

NeuralLinear(d::Int64, num_arms::Int64, layer_sizes::Vector{Int64}, inital_batches::Int64,
             retrain_freq::Int64, epochs::Int64)

NeuralLinear policy introduced in the paper Deep Bayesian Bandits Showdown: An Empirical Comparison of Bayesian Deep Networks for Thompson Sampling.

Keyword Arguments

opt=ADAM(): Optimizer used to update the neural network parameters.

λ::Float64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

verbose_retrain::Bool=false: Print the details of the neural network training procedure.

NonlinearBandits.fit! — Method

fit!(enc::NeuralEncoder, X::AbstractMatrix, a::AbstractVector{<:Int}, r::AbstractMatrix,
     epochs::Int64)

Update a the neural network parameters using a contextual multi-armed bandit trajectory.

Keyword Arguments

batch_size::Int64=32: The batch size to use while training the neural network.

opt=ADAM(): The optimizer to use for updating the parameters.

verbose::Bool=true: Print details of the fitting procedure.

Reward Samplers

NonlinearBandits.GaussianRewards — Type

GaussianRewards(mf::Tuple{Vararg{<:Function}}; <keyword arguments>)

Construct a callable object to sample gaussian rewards.

Arguments

mf::Tuple{Vararg{<:Function}}: A Tuple of functions which take a 1-dimensional input and output the (scalar) mean reward for the corresponding action.

σ::Float64: The standard deviation of the gaussian noise applied to each reward.

Metrics

NonlinearBandits.FunctionalRegret — Type

FunctionalRegret(mf::Tuple{Vararg{<:Function}})

Metric to track the regret of each action using a discrete set of functions.

Drivers

NonlinearBandits.StandardDriver — Type

StandardDriver(csampler::AbstractContextSampler, policy::AbstractPolicy, 
               rsampler::AbstractRewardSampler[, 
               metrics::Tuple{Vararg{<:AbstractMetric}}])

A simple driver that samples contexts, passes them to the policy to generate actions, then observes the rewards.

Arguments

csampler::AbstractContextSampler: Context sampler.

policy::AbstractPolicy: Policy to generate actions given contexts.

rsampler::AbstractRewardSampler: Sampler for rewards, given the contexts and actions.

metrics::Tuple{Vararg{<:AbstractMetric}}: A tuple of metrics that will each be called as metric(X, a, r).

NonlinearBandits.run! — Method

run!(num_batches::Int64, batch_size::Int64, driver::AbstractDriver; <keyword arguments>)

Run a driver for num_batches batches.