Model API

NonlinearBandits.fit! — Method

fit!(model, X::AbstractMatrix, y::AbstractMatrix)

Update the parameters of model.

Arguments

X::AbstractMatrix: A matrix with observations stored as columns.

y::AbstractMatrix: A matrix with 1 row of response variables.

Polynomials

NonlinearBandits.AbstractBayesianLM — Type

Abstract type for models using a Gaussian/normal-inverse-gamma conjugate prior.

NonlinearBandits.BayesLM — Type

BayesLM(d::Int; <keyword arguments>)

Construct a Bayesian linear model.

Arguments

λ::Float64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

NonlinearBandits.std — Method

std(model::AbstractBayesianLM)

Return the posterior mean of the inverse-gamma distribution.

NonlinearBandits.BayesPM — Type

BayesPM(basis::Vector{Index}, limits::Matrix{Float64}; <keyword arguments>)

Construct a Bayesian linear model on polynomial features.

Arguments

basis::Vector{Index}: Vector of monomial indices.

limits::Matrix{Float64}: Matrix with two columns defining the lower/upper limits of the space.

λ::Float64=1.0: Prior covariance scale factor.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e=3: Inverse-gamma prior scale hyperparameter.

NonlinearBandits.Index — Type

Index(dim::Vector{Int64}, deg::Vector{Int64})

Multivariate monomial index.

The monomial x[1] * x[3]^2can be encoded usingdim = [1, 3],deg = [1, 2]`

NonlinearBandits.Partition — Type

Partition(limits::Matrix{Float64})

Construct an object capable of storing a hyperrectangular partition.

NonlinearBandits.PartitionedBayesPM — Method

PartitionedBayesPM(X::AbstractMatrix, y::AbstractMatrix, limits::Matrix{Float64};
                   <keyword arguments>)

Perform a 1-step look ahead greedy search for a partitioned polynomial model.

Keyword Arguments

Jmax::Int64=3: The maximum degree of any polynomial model.

Pmax::Int64=500: The maximum number of features in a particular regions.

Kmax::Int64=200: The maximum number of regions

λ::Float64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

ratio::Float64=1.0: Polynomial degrees are reduced until size(X, 2) < ratio * length(tpbasis(d, J)).

tol::Float64=1e-4: The required increase in the model evidence to accept a split.

verbose::Bool=true: Print details of the partition search.

NonlinearBandits.PartitionedBayesPM — Method

PartitionedBayesPM(P::Partition, Js::Vector{Int64}; <keyword arguments>)

Contruct a partitioned polynomial model.

Arguments

P::Partition: A partition of the space.

λ::Float64=1.0: Prior scaling.

shape0::Float64=1e-3: Inverse-gamma prior shape hyperparameter.

scale0::Float64=1e-3: Inverse-gamma prior scale hyperparameter.

NonlinearBandits.expand — Method

expand(X::AbstractMatrix, basis::Vector{Index}, limits::Matrix{Float64};
       J::Union{Nothing,Int64}=nothing)

Expand the columns of X into a rescaled legendre polynomial basis.

Arguments

X::AbstractMatrix: Matrix with observations stored as columns.

basis::Vector{Index}: Vector of monomial indices.

limits::Matrix{Float64}: Matrix with two columns defining the lower/upper limits of the space.

J::Union{Nothing, Int64}=nothing: The maximum degree of the basis. Inferred if not specified.

NonlinearBandits.lasso_selection — Method

lasso_selection(X::AbstractMatrix, y::AbstractMatrix, Pmax::Int64, intercept::Bool)

Choose the first Pmax features introduced by a LASSO solution path.

Arguments

X::AbstractMatrix: A matrix with observations stored as columns.

y::AbstractMatrix: A matrix with 1 row of response variables.

Pmax::Int64: The maximum number of predictors.

intercept::Bool: true if the first row of X are the intercept features

NonlinearBandits.locate — Method

locate(P::Partition, X::AbstractMatrix)

Return a vector of integers giving the region index for each column of X.

NonlinearBandits.split! — Method

split!(P::Partition, k::Int64, d::Int64)

Split the k'th subregion of P into equal halves in dimension d.

NonlinearBandits.tpbasis — Method

tpbasis(d::Int64, J::Int64)

Construct the d-dimensional truncated tensor-product basis.

All index terms have a degree ≤ J.

See also Index

NonlinearBandits.shape_scale — Method

shape_scale(model::AbstractBayesianLM)

Return the shape/scale of model.