NonlinearBandits

Documentation for NonlinearBandits.

NonlinearBandits.BayesLM

NonlinearBandits.fit!

NonlinearBandits.std

NonlinearBandits.BayesLM — Method

BayesLM(d::Int; λ::AbstractFloat=1.0, shape0::AbsractFloat=1e-3,
        scale0::AbstractFloat=1e-3)

Construct a Bayesian linear model with prior scaling λ.

NonlinearBandits.fit! — Method

fit!(model, X::AbstractMatrix, y::AbstractMatrix)

Update the parameters of model.

Arguments

X::AbstractMatrix: A matrix with observations stored as columns.

y::AbstractMatrix: A matrix with 1 row of response variables.

NonlinearBandits.std — Method

std(lm::BayesLM)

Computed the posterior expected standard deviation from the model lm.